requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
cryptography==41.0.5
python-restcountries==2.0.0
pyjwt
python-dotenv==1.0.0
pydantic[email]
pymysql==1.1.0
python-logstash-async